Add a LodePNG-based image loader to preserve my sanity
Allegro-based image loader doesn't like multi-threaded image loading - at
least not from threads that are not created by Allegro itself. It offloads
loading some of the formats to the system, which is generally good, since
it doesn't require an additional image-loading library, but in case of
Android and multithreading this causes a load in a thread that might not
have its JNIEnv pointer - and thus fails.

tl;dr: don't trust Allegro's image loader.